Split out the Github plugin methods
Split the various chunks of the Github API controller module into logical,
smaller, chunks. It doesn't make the plugin any less complex, but it makes it
easier to read.
