+++ app/cyclid/plugins/api/github/config.rb
@@ -1,0 +1,45 @@
+# frozen_string_literal: true
+# Copyright 2016 Liqwyd Ltd.
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#     http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+
+require_relative 'oauth'
+
+# Top level module for the core Cyclid code.
+module Cyclid
+  # Module for the Cyclid API
+  module API
+    # Module for Cyclid Plugins
+    module Plugins
+      # Container for the Sinatra related controllers modules
+      module ApiExtension
+        # Github plugin method callbacks
+        module GithubMethods
+          # Plugin configuration methods
+          module Config
+            # Load the config for the Github plugin and set defaults if they're not
+            # in the config
+            def load_github_config(config)
+              config.symbolize_keys!
+
+              server_config = config[:github] || {}
+              Cyclid.logger.debug "config=#{server_config}"
+
+              server_config.symbolize_keys
+            end
+          end
+        end
+      end
+    end
+  end
+end
